Add files via upload
Task 1: task is to implement a novel data structure - your project lead is calling it a power of two max heap. 
The heap must satisfy the heap property.
Every parent node in the heap must have 2^x children.
The value of x must be a parameter of the heap’s constructor.
The heap must implement an insert method.
The heap must implement a pop max method.
The heap must be implemented in Java.
The heap must be performant.
You must use a more descriptive variable name than x in your implementation.